custom-forms-app/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Werkzeug==2.3.7
python-dotenv==1.0.0
email_validator==2.0.0
openpyxl==3.1.2
Pillow==10.0.0
reportlab==4.0.0
